Rough implementation of collision detection for the background
Generate a collision map from collision.png. Each byte represents 8 horizontal tiles, so each row takes up 4 bytes. Each row represents a single vertical block, of which there are 15, so the entire map takes up 60 bytes.

This implementation only looks whether the byte is 0 or not, which effectively means collision detection is only granular to each quarter of the screen, horizontally. In the future this will be improved so that collision detection is granular to a horizontal tile, and a vertical block. This change will involve looking at individual bits within the collision map.