Fly swatter is immortal, and boomerangs towards player in the h dimension.
Instead of having a limited lifetime, and a constant speed for fly swatter, set an initial speed, and modify that speed every frame, based upon the difference from the player. In order to have this work, add a high byte to the object position, and a low byte to the object speed. The object position high byte handles wraps across screens. The object speed low byte allows for smooth acceleration.